backend/requirements.txt
django
djangorestframework
django-rest-knox
django_countries
django-cleanup
django-cors-headers
Pillow
django-notifications-hq